Tested. Need to remove temporary logging, and need to return actual PWM frequency during init.

+++ FQuad/Platform/PlatformPWM/PlatformPWM.h
@@ -24,7 +24,7 @@ typedef enum
 
 PlatformStatus PlatformPWM_Init( const PlatformPWM_t inPWM, const uint32_t inRequestedPWMFrequency );
 
-PlatformStatus PlatformPWM_Start( const PlatformPWM_t inPWM, const uint8_t inDutyCycle );
+PlatformStatus PlatformPWM_Start( const PlatformPWM_t inPWM, const float inDutyCycle );
 
 PlatformStatus PlatformPWM_Stop( const PlatformPWM_t inPWM );
